requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
phonenumbers==8.12.20
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
